FDG-9761 download disable works
