+++ src/components/data-preview/data-preview-filter-section/data-preview-download/data-preview-download-button/data-preview-download-button.module.scss
@@ -23,12 +23,14 @@ $icon-width: 1.5rem;
   &.buttonActive {
     background-color: #112e51;
   }
+  &:disabled {
+    opacity: 40%;
+  }
 }
 
 .parent {
   position: relative;
 }
-
 
 @media screen and (min-width: $breakpoint-lg) {
   .downloadButton {
@@ -45,6 +47,5 @@ $icon-width: 1.5rem;
       width: calc(100% - $icon-width);
       margin-left: $icon-width;
     }
-
   }
 }
