+++ src/components/data-preview/data-preview-filter-section/data-preview-download/data-preview-download-button/data-preview-download-button.tsx
@@ -1,4 +1,4 @@
-import React, { FunctionComponent } from 'react';
+import React, { FunctionComponent, useEffect, useState } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCaretDown, faCaretUp, faCaretRight, faCloudDownload } from '@fortawesome/free-solid-svg-icons';
 import { pxToNumber } from '../../../../../helpers/styles-helper/styles-helper';
@@ -13,7 +13,8 @@ interface IDownloadButtonProps {
   disabled: boolean;
 }
 
-const DataPreviewDownloadButton: FunctionComponent<IDownloadButtonProps> = ({ active, setActive, width, disabled }: IDownloadButtonProps) => {
+const DataPreviewDownloadButton: FunctionComponent<IDownloadButtonProps> = ({ active, setActive, width, apiFilterDefault }: IDownloadButtonProps) => {
+  const [isDisabled, setIsDisabled] = useState(false);
   const getIcon = desktopWidth => {
     if (desktopWidth) {
       return active ? faCaretUp : faCaretDown;
@@ -21,15 +22,25 @@ const DataPreviewDownloadButton: FunctionComponent<IDownloadButtonProps> = ({ ac
       return active ? faCloudDownload : faCaretRight;
     }
   };
+  console.log(apiFilterDefault);
+  const toggleDisabled = () => {
+    if (apiFilterDefault === true) {
+      setIsDisabled(true);
+    }
+  };
   const activateToggle = () => {
-    if (disabled !== true) {
+    if (isDisabled !== true) {
       setActive(!active);
     }
   };
 
+  useEffect(() => {
+    toggleDisabled();
+  }, [apiFilterDefault]);
+
   return (
     <div className={parent}>
-      <button className={`${downloadButton} ${active && buttonActive}`} onClick={() => activateToggle()}>
+      <button className={`${downloadButton} ${active && buttonActive}`} onClick={() => activateToggle()} disabled={isDisabled}>
         <div className={buttonText}>Download</div>
         <div className={icon}>
           <FontAwesomeIcon icon={getIcon(width >= pxToNumber(breakpointLg))} />
